Move all the local stuff to the local package
