+++ nextmv/nextmv/__init__.py
@@ -16,6 +16,14 @@ from .input import text_data_file as text_data_file
 from .logger import log as log
 from .logger import redirect_stdout as redirect_stdout
 from .logger import reset_stdout as reset_stdout
+from .manifest import MANIFEST_FILE_NAME as MANIFEST_FILE_NAME
+from .manifest import Manifest as Manifest
+from .manifest import ManifestBuild as ManifestBuild
+from .manifest import ManifestOption as ManifestOption
+from .manifest import ManifestPython as ManifestPython
+from .manifest import ManifestPythonModel as ManifestPythonModel
+from .manifest import ManifestRuntime as ManifestRuntime
+from .manifest import ManifestType as ManifestType
 from .model import Model as Model
 from .model import ModelConfiguration as ModelConfiguration
 from .options import Option as Option
@@ -40,6 +48,29 @@ from .output import json_solution_file as json_solution_file
 from .output import text_solution_file as text_solution_file
 from .output import write as write
 from .output import write_local as write_local
+from .polling import DEFAULT_POLLING_OPTIONS as DEFAULT_POLLING_OPTIONS
+from .polling import PollingOptions as PollingOptions
+from .polling import poll as poll
+from .run import ErrorLog as ErrorLog
+from .run import ExternalRunResult as ExternalRunResult
+from .run import Format as Format
+from .run import FormatInput as FormatInput
+from .run import FormatOutput as FormatOutput
+from .run import Metadata as Metadata
+from .run import RunConfiguration as RunConfiguration
+from .run import RunInformation as RunInformation
+from .run import RunLog as RunLog
+from .run import RunQueuing as RunQueuing
+from .run import RunResult as RunResult
+from .run import RunType as RunType
+from .run import RunTypeConfiguration as RunTypeConfiguration
+from .run import TrackedRun as TrackedRun
+from .run import TrackedRunStatus as TrackedRunStatus
+from .run import run_duration as run_duration
+from .safe import safe_id as safe_id
+from .safe import safe_name_and_id as safe_name_and_id
+from .status import Status as Status
+from .status import StatusV2 as StatusV2
 
 VERSION = __version__
 """The version of the Nextmv Python SDK."""
