+++ nextmv/nextmv/local/__init__.py
@@ -1,3 +1,5 @@
 """
 Functionality for locally simulating the Nextmv Cloud.
 """
+
+from .application import Application as Application
