+++ nextmv/nextmv/local/geojson_handler.py
