+++ nextmv/tests/local/test_application.py
@@ -1,41 +1,21 @@
+import json
 import os
+import shutil
 import tempfile
 import unittest
-from typing import Any
 from unittest.mock import Mock, patch
 
-from nextmv.cloud.application import _DEFAULT_POLLING_OPTIONS, Application, PollingOptions, poll
-from nextmv.cloud.client import Client
-from nextmv.cloud.run import Format, FormatInput, FormatOutput, RunConfiguration
-from nextmv.cloud.status import StatusV2
-from nextmv.input import InputFormat
+import yaml
+from nextmv.input import Input, InputFormat
+from nextmv.local.application import Application
+from nextmv.options import Options
 from nextmv.output import OutputFormat
-
-
-# This is a dummy function to avoid actually sleeping during tests.
-def no_sleep(value: float) -> None:
-    return
+from nextmv.polling import DEFAULT_POLLING_OPTIONS, PollingOptions
+from nextmv.run import Format, FormatInput, FormatOutput, RunConfiguration, RunInformation, RunResult
+from nextmv.status import StatusV2
 
 
 class TestApplication(unittest.TestCase):
-    def test_poll(self):
-        counter = 0
-
-        def polling_func() -> tuple[Any, bool]:
-            nonlocal counter
-            counter += 1
-
-            if counter < 4:
-                return "result", False
-
-            return "result", True
-
-        polling_options = PollingOptions()
-
-        result = poll(polling_options, polling_func, no_sleep)
-
-        self.assertEqual(result, "result")
-
     def test_initialize(self):
         """Test the Application.initialize method."""
         with tempfile.TemporaryDirectory() as temp_dir:
@@ -43,21 +23,16 @@ class TestApplication(unittest.TestCase):
             app_id = "test-app-id"
             description = "Test application"
 
-            # Mock client
-            mock_client = Mock(spec=Client)
-
             # Initialize the application
             app = Application.initialize(
                 name=app_name,
                 id=app_id,
                 description=description,
                 destination=temp_dir,
-                client=mock_client,
             )
 
             # Verify the application object
             self.assertEqual(app.id, app_id)
-            self.assertEqual(app.client, mock_client)
             self.assertEqual(app.description, description)
             self.assertEqual(app.src, os.path.join(temp_dir, app_name))
 
@@ -98,7 +73,6 @@ class TestApplication(unittest.TestCase):
 
                 # Verify the application object has generated ID
                 self.assertIsNotNone(app.id)
-                self.assertIsNone(app.client)
                 self.assertIsNone(app.description)  # description should be None when not provided
                 # Use the current working directory for comparison since that's where the app is created
                 expected_src_path = os.path.join(os.getcwd(), app_name)
@@ -138,55 +112,9 @@ class TestApplication(unittest.TestCase):
             self.assertTrue(os.path.exists(app_dir))
             self.assertTrue(os.path.exists(os.path.join(app_dir, "app.yaml")))
 
-    def test_poll_stop_callback(self):
-        counter = 0
-
-        # The polling func would stop after 9 calls.
-        def polling_func() -> tuple[Any, bool]:
-            nonlocal counter
-            counter += 1
-
-            if counter < 10:
-                return "result", False
-
-            return "result", True
-
-        # The stop callback makes sure that the polling stops sooner, after 3
-        # calls.
-        def stop() -> bool:
-            if counter == 3:
-                return True
-
-        polling_options = PollingOptions(stop=stop)
-
-        result = poll(polling_options, polling_func, no_sleep)
-
-        self.assertIsNone(result)
-
-    def test_poll_long(self):
-        counter = 0
-        max_tries = 1000000
-
-        def polling_func() -> tuple[Any, bool]:
-            nonlocal counter
-            counter += 1
-
-            if counter < max_tries:
-                return "result", False
-
-            return "result", True
-
-        polling_options = PollingOptions(
-            max_tries=max_tries + 1,
-        )
-
-        result = poll(polling_options, polling_func, no_sleep)
-
-        self.assertEqual(result, "result")
-
 
 class TestApplicationNewLocalRun(unittest.TestCase):
-    """Test cases for the Application.new_local_run method."""
+    """Test cases for the Application.new_run method."""
 
     def setUp(self):
         """Set up test fixtures."""
@@ -207,8 +135,6 @@ class TestApplicationNewLocalRun(unittest.TestCase):
         }
 
         with open(os.path.join(self.app_src, "app.yaml"), "w") as f:
-            import yaml
-
             yaml.dump(manifest_content, f)
 
         # Create a simple entrypoint
@@ -227,20 +153,18 @@ print(json.dumps(output))
 """)
 
         # Create test application
-        self.app = Application(id="test-app", client=Mock(spec=Client), src=self.app_src)
+        self.app = Application(id="test-app", src=self.app_src)
 
     def tearDown(self):
         """Clean up test fixtures."""
-        import shutil
-
         shutil.rmtree(self.test_dir, ignore_errors=True)
 
-    @patch("nextmv.cloud.application.run")
-    def test_new_local_run_basic(self, mock_run):
-        """Test basic new_local_run functionality."""
+    @patch("nextmv.local.application.run")
+    def test_new_run_basic(self, mock_run):
+        """Test basic new_run functionality."""
         mock_run.return_value = "test-run-id"
 
-        result = self.app.new_local_run(input={"test": "input"}, options={"duration": "10s"})
+        result = self.app.new_run(input={"test": "input"}, options={"duration": "10s"})
 
         self.assertEqual(result, "test-run-id")
         mock_run.assert_called_once()
@@ -253,9 +177,9 @@ print(json.dumps(output))
         self.assertEqual(call_args[1]["input_data"], {"test": "input"})
         self.assertEqual(call_args[1]["options"], {"duration": "10s"})
 
-    @patch("nextmv.cloud.application.run")
-    def test_new_local_run_with_inputs_dir_path(self, mock_run):
-        """Test new_local_run with inputs directory path."""
+    @patch("nextmv.local.application.run")
+    def test_new_run_with_inputs_dir_path(self, mock_run):
+        """Test new_run with inputs directory path."""
         mock_run.return_value = "test-run-id-2"
 
         # Create test inputs directory
@@ -264,7 +188,7 @@ print(json.dumps(output))
         with open(os.path.join(inputs_dir, "data.csv"), "w") as f:
             f.write("col1,col2\nval1,val2\n")
 
-        result = self.app.new_local_run(
+        result = self.app.new_run(
             input_dir_path=inputs_dir,
             configuration=RunConfiguration(
                 format=Format(
@@ -281,14 +205,14 @@ print(json.dumps(output))
         call_args = mock_run.call_args
         self.assertEqual(call_args[1]["inputs_dir_path"], inputs_dir)
 
-    @patch("nextmv.cloud.application.run")
-    def test_new_local_run_with_dict_input(self, mock_run):
-        """Test new_local_run with dictionary input."""
+    @patch("nextmv.local.application.run")
+    def test_new_run_with_dict_input(self, mock_run):
+        """Test new_run with dictionary input."""
         mock_run.return_value = "test-run-id-3"
 
         input_data = {"vehicles": [{"id": 1, "capacity": 100}]}
 
-        result = self.app.new_local_run(input=input_data)
+        result = self.app.new_run(input=input_data)
 
         self.assertEqual(result, "test-run-id-3")
 
@@ -296,14 +220,14 @@ print(json.dumps(output))
         call_args = mock_run.call_args
         self.assertEqual(call_args[1]["input_data"], input_data)
 
-    @patch("nextmv.cloud.application.run")
-    def test_new_local_run_with_string_input(self, mock_run):
-        """Test new_local_run with string input."""
+    @patch("nextmv.local.application.run")
+    def test_new_run_with_string_input(self, mock_run):
+        """Test new_run with string input."""
         mock_run.return_value = "test-run-id-4"
 
         input_data = "raw text input"
 
-        result = self.app.new_local_run(input=input_data)
+        result = self.app.new_run(input=input_data)
 
         self.assertEqual(result, "test-run-id-4")
 
@@ -311,42 +235,31 @@ print(json.dumps(output))
         call_args = mock_run.call_args
         self.assertEqual(call_args[1]["input_data"], input_data)
 
-    def test_new_local_run_no_src_error(self):
-        """Test new_local_run raises error when src is not set."""
-        app = Application(id="test-app", client=Mock(spec=Client))
-
+    def test_new_run_no_input_error(self):
+        """Test new_run raises error when neither input nor inputs_dir_path is provided."""
         with self.assertRaises(ValueError) as context:
-            app.new_local_run(input={"test": "data"})
-
-        self.assertIn("`src` property for the `Application` must be specified", str(context.exception))
-
-    def test_new_local_run_no_input_error(self):
-        """Test new_local_run raises error when neither input nor inputs_dir_path is provided."""
-        with self.assertRaises(ValueError) as context:
-            self.app.new_local_run()
+            self.app.new_run()
 
         self.assertIn("Either `input` or `input_directory` must be specified", str(context.exception))
 
-    def test_new_local_run_manifest_not_found_error(self):
-        """Test new_local_run raises error when manifest.yaml is not found."""
+    def test_new_run_manifest_not_found_error(self):
+        """Test new_run raises error when manifest.yaml is not found."""
         # Remove the manifest file
         os.remove(os.path.join(self.app_src, "app.yaml"))
 
         with self.assertRaises(FileNotFoundError) as context:
-            self.app.new_local_run(input={"test": "data"})
+            self.app.new_run(input={"test": "data"})
 
         self.assertIn("Could not find manifest.yaml", str(context.exception))
 
-    @patch("nextmv.cloud.application.run")
-    def test_new_local_run_with_nextmv_input_object(self, mock_run):
-        """Test new_local_run with nextmv.Input object."""
+    @patch("nextmv.local.application.run")
+    def test_new_run_with_nextmv_input_object(self, mock_run):
+        """Test new_run with nextmv.Input object."""
         mock_run.return_value = "test-run-id-5"
-
-        from nextmv import Input
 
         input_obj = Input(data={"test": "input_object"})
 
-        result = self.app.new_local_run(input=input_obj)
+        result = self.app.new_run(input=input_obj)
 
         self.assertEqual(result, "test-run-id-5")
 
@@ -354,18 +267,16 @@ print(json.dumps(output))
         call_args = mock_run.call_args
         self.assertEqual(call_args[1]["input_data"], {"test": "input_object"})
 
-    @patch("nextmv.cloud.application.run")
-    def test_new_local_run_with_options_object(self, mock_run):
-        """Test new_local_run with nextmv.Options object."""
+    @patch("nextmv.local.application.run")
+    def test_new_run_with_options_object(self, mock_run):
+        """Test new_run with nextmv.Options object."""
         mock_run.return_value = "test-run-id-6"
-
-        from nextmv import Options
 
         options_obj = Options()
         options_obj.duration = "30s"
         options_obj.iterations = 500
 
-        result = self.app.new_local_run(input={"test": "data"}, options=options_obj)
+        result = self.app.new_run(input={"test": "data"}, options=options_obj)
 
         self.assertEqual(result, "test-run-id-6")
 
@@ -374,14 +285,10 @@ print(json.dumps(output))
         expected_options = {"duration": "30s", "iterations": "500"}
         self.assertEqual(call_args[1]["options"], expected_options)
 
-    @patch("nextmv.cloud.application.run")
-    def test_new_local_run_with_configuration_object(self, mock_run):
-        """Test new_local_run with RunConfiguration object."""
+    @patch("nextmv.local.application.run")
+    def test_new_run_with_configuration_object(self, mock_run):
+        """Test new_run with RunConfiguration object."""
         mock_run.return_value = "test-run-id-7"
-
-        from nextmv.cloud.run import Format, FormatInput, FormatOutput, RunConfiguration
-        from nextmv.input import InputFormat
-        from nextmv.output import OutputFormat
 
         config = RunConfiguration(
             format=Format(
@@ -390,7 +297,7 @@ print(json.dumps(output))
             )
         )
 
-        result = self.app.new_local_run(input={"test": "data"}, configuration=config)
+        result = self.app.new_run(input={"test": "data"}, configuration=config)
 
         self.assertEqual(result, "test-run-id-7")
 
@@ -400,8 +307,8 @@ print(json.dumps(output))
         run_config = call_args[1]["run_config"]
         self.assertEqual(run_config["format"]["input"]["type"], "json")
 
-    @patch("nextmv.cloud.application.run")
-    def test_new_local_run_inputs_dir_path_takes_precedence(self, mock_run):
+    @patch("nextmv.local.application.run")
+    def test_new_run_inputs_dir_path_takes_precedence(self, mock_run):
         """Test that inputs_dir_path takes precedence over input."""
         mock_run.return_value = "test-run-id-8"
 
@@ -411,7 +318,7 @@ print(json.dumps(output))
         with open(os.path.join(inputs_dir, "test.txt"), "w") as f:
             f.write("file content")
 
-        result = self.app.new_local_run(
+        result = self.app.new_run(
             input={"should": "be ignored"},
             input_dir_path=inputs_dir,
             configuration=RunConfiguration(
@@ -429,19 +336,19 @@ print(json.dumps(output))
         self.assertIsNone(call_args[1]["input_data"])
         self.assertEqual(call_args[1]["inputs_dir_path"], inputs_dir)
 
-    @patch("nextmv.cloud.application.run")
-    def test_new_local_run_json_configurations(self, mock_run):
-        """Test new_local_run with json_configurations parameter."""
+    @patch("nextmv.local.application.run")
+    def test_new_run_json_configurations(self, mock_run):
+        """Test new_run with json_configurations parameter."""
         mock_run.return_value = "test-run-id-9"
 
         json_configs = {"ensure_ascii": False, "indent": 2}
 
-        result = self.app.new_local_run(input={"test": "data"}, json_configurations=json_configs)
+        result = self.app.new_run(input={"test": "data"}, json_configurations=json_configs)
 
         self.assertEqual(result, "test-run-id-9")
         mock_run.assert_called_once()
 
-    def test_new_local_run_validate_dir_path_and_configuration(self):
+    def test_new_run_validate_dir_path_and_configuration(self):
         """Test validation of inputs_dir_path and configuration parameters."""
         # Create test inputs directory
         inputs_dir = os.path.join(self.test_dir, "inputs")
@@ -449,13 +356,13 @@ print(json.dumps(output))
 
         # Should raise error when inputs_dir_path is provided without configuration
         with self.assertRaises(ValueError):
-            self.app.new_local_run(input_dir_path=inputs_dir)
+            self.app.new_run(input_dir_path=inputs_dir)
 
         # Should work when both are provided
-        with patch("nextmv.cloud.application.run") as mock_run:
+        with patch("nextmv.local.application.run") as mock_run:
             mock_run.return_value = "test-run-id"
 
-            result = self.app.new_local_run(
+            result = self.app.new_run(
                 input_dir_path=inputs_dir,
                 configuration=RunConfiguration(
                     format=Format(
@@ -469,8 +376,8 @@ print(json.dumps(output))
 
 
 class TestApplicationLocalRunMethods(unittest.TestCase):
-    """Test cases for the Application local run methods: local_run_metadata, local_run_result,
-    local_run_result_with_polling, and new_local_run_with_result."""
+    """Test cases for the Application local run methods: run_metadata, run_result,
+    run_result_with_polling, and new_run_with_result."""
 
     def setUp(self):
         """Set up test fixtures."""
@@ -483,7 +390,7 @@ class TestApplicationLocalRunMethods(unittest.TestCase):
         os.makedirs(self.runs_dir)
 
         # Create test application
-        self.app = Application(id="test-app", client=Mock(spec=Client), src=self.app_src)
+        self.app = Application(id="test-app", src=self.app_src)
 
         # Test run ID
         self.test_run_id = "run-123"
@@ -503,14 +410,10 @@ class TestApplicationLocalRunMethods(unittest.TestCase):
         }
 
         with open(os.path.join(self.app_src, "app.yaml"), "w") as f:
-            import yaml
-
             yaml.dump(manifest_content, f)
 
     def tearDown(self):
         """Clean up test fixtures."""
-        import shutil
-
         shutil.rmtree(self.test_dir, ignore_errors=True)
 
     def _create_run_info_file(self, status="succeeded", output_type="json", error=""):
@@ -539,8 +442,6 @@ class TestApplicationLocalRunMethods(unittest.TestCase):
 
         info_file = os.path.join(self.test_run_dir, f"{self.test_run_id}.json")
         with open(info_file, "w") as f:
-            import json
-
             json.dump(run_info, f)
 
     def _create_output_files(self, output_type="JSON", output_data=None):
@@ -551,121 +452,112 @@ class TestApplicationLocalRunMethods(unittest.TestCase):
         if output_type == "JSON":
             solution_file = os.path.join(outputs_dir, "solution.json")
             with open(solution_file, "w") as f:
-                import json
-
                 json.dump(output_data or {"solution": {"result": 42}}, f)
         elif output_type in ["CSV_ARCHIVE", "MULTI_FILE"]:
             # Create some dummy files for non-JSON outputs
             with open(os.path.join(outputs_dir, "result.csv"), "w") as f:
                 f.write("id,value\n1,42\n")
 
-    # Tests for local_run_metadata
-    def test_local_run_metadata_success(self):
-        """Test successful local_run_metadata call."""
-        from nextmv.cloud.run import RunInformation
+    # Tests for run_metadata
+    def test_run_metadata_success(self):
+        """Test successful run_metadata call."""
 
         self._create_run_info_file(status="succeeded")
 
-        result = self.app.local_run_metadata(self.test_run_id)
+        result = self.app.run_metadata(self.test_run_id)
 
         self.assertIsInstance(result, RunInformation)
         self.assertEqual(result.id, self.test_run_id)
         self.assertEqual(result.metadata.status_v2, "succeeded")
 
-    def test_local_run_metadata_no_runs_dir(self):
-        """Test local_run_metadata when .nextmv/runs directory doesn't exist."""
+    def test_run_metadata_no_runs_dir(self):
+        """Test run_metadata when .nextmv/runs directory doesn't exist."""
         # Remove the runs directory
-        import shutil
-
         shutil.rmtree(self.runs_dir)
 
         with self.assertRaises(ValueError) as context:
-            self.app.local_run_metadata(self.test_run_id)
+            self.app.run_metadata(self.test_run_id)
 
         self.assertIn("`.nextmv/runs` dir does not exist", str(context.exception))
 
-    def test_local_run_metadata_no_run_dir(self):
-        """Test local_run_metadata when specific run directory doesn't exist."""
+    def test_run_metadata_no_run_dir(self):
+        """Test run_metadata when specific run directory doesn't exist."""
         with self.assertRaises(ValueError) as context:
-            self.app.local_run_metadata("non-existent-run")
+            self.app.run_metadata("non-existent-run")
 
         self.assertIn("run dir does not exist", str(context.exception))
 
-    def test_local_run_metadata_no_info_file(self):
-        """Test local_run_metadata when run info file doesn't exist."""
+    def test_run_metadata_no_info_file(self):
+        """Test run_metadata when run info file doesn't exist."""
         # Create run directory but no info file
         non_existent_run_id = "run-456"
         os.makedirs(os.path.join(self.runs_dir, non_existent_run_id))
 
         with self.assertRaises(ValueError) as context:
-            self.app.local_run_metadata(non_existent_run_id)
+            self.app.run_metadata(non_existent_run_id)
 
         self.assertIn("file does not exist", str(context.exception))
 
-    def test_local_run_metadata_failed_status(self):
-        """Test local_run_metadata with failed status."""
-        from nextmv.cloud.run import RunInformation
+    def test_run_metadata_failed_status(self):
+        """Test run_metadata with failed status."""
 
         self._create_run_info_file(status="failed", error="Test error message")
 
-        result = self.app.local_run_metadata(self.test_run_id)
+        result = self.app.run_metadata(self.test_run_id)
 
         self.assertIsInstance(result, RunInformation)
         self.assertEqual(result.metadata.status_v2, "failed")
         self.assertEqual(result.metadata.error, "Test error message")
 
-    # Tests for local_run_result
-    @patch.object(Application, "local_run_metadata")
-    @patch.object(Application, "_Application__local_run_result")
-    def test_local_run_result_success(self, mock_local_run_result, mock_local_run_metadata):
-        """Test successful local_run_result call."""
-        from nextmv.cloud.run import RunInformation, RunResult
+    # Tests for run_result
+    @patch.object(Application, "run_metadata")
+    @patch.object(Application, "_Application__run_result")
+    def test_run_result_success(self, mock_run_result, mock_run_metadata):
+        """Test successful run_result call."""
 
         # Mock the metadata response
         mock_run_info = Mock(spec=RunInformation)
-        mock_local_run_metadata.return_value = mock_run_info
+        mock_run_metadata.return_value = mock_run_info
 
         # Mock the result response
         mock_result = Mock(spec=RunResult)
-        mock_local_run_result.return_value = mock_result
+        mock_run_result.return_value = mock_result
 
-        result = self.app.local_run_result(self.test_run_id, output_dir_path="/tmp")
+        result = self.app.run_result(self.test_run_id, output_dir_path="/tmp")
 
         # Verify calls
-        mock_local_run_metadata.assert_called_once_with(run_id=self.test_run_id)
-        mock_local_run_result.assert_called_once_with(
+        mock_run_metadata.assert_called_once_with(run_id=self.test_run_id)
+        mock_run_result.assert_called_once_with(
             run_id=self.test_run_id, run_information=mock_run_info, output_dir_path="/tmp"
         )
         self.assertEqual(result, mock_result)
 
-    @patch.object(Application, "local_run_metadata")
-    @patch.object(Application, "_Application__local_run_result")
-    def test_local_run_result_default_output_dir(self, mock_local_run_result, mock_local_run_metadata):
-        """Test local_run_result with default output directory."""
-        from nextmv.cloud.run import RunInformation, RunResult
+    @patch.object(Application, "run_metadata")
+    @patch.object(Application, "_Application__run_result")
+    def test_run_result_default_output_dir(self, mock_run_result, mock_run_metadata):
+        """Test run_result with default output directory."""
 
         # Mock the responses
         mock_run_info = Mock(spec=RunInformation)
-        mock_local_run_metadata.return_value = mock_run_info
+        mock_run_metadata.return_value = mock_run_info
         mock_result = Mock(spec=RunResult)
-        mock_local_run_result.return_value = mock_result
+        mock_run_result.return_value = mock_result
 
-        result = self.app.local_run_result(self.test_run_id)
+        result = self.app.run_result(self.test_run_id)
 
         # Verify default output_dir_path is used
-        mock_local_run_result.assert_called_once_with(
+        mock_run_result.assert_called_once_with(
             run_id=self.test_run_id, run_information=mock_run_info, output_dir_path="."
         )
 
         self.assertEqual(result, mock_result)
 
-    # Tests for local_run_result_with_polling
-    @patch.object(Application, "local_run_metadata")
-    @patch.object(Application, "_Application__local_run_result")
-    @patch("nextmv.cloud.application.poll")
-    def test_local_run_result_with_polling_success(self, mock_poll, mock_local_run_result, mock_local_run_metadata):
-        """Test successful local_run_result_with_polling call."""
-        from nextmv.cloud.run import RunInformation, RunResult
+    # Tests for run_result_with_polling
+    @patch.object(Application, "run_metadata")
+    @patch.object(Application, "_Application__run_result")
+    @patch("nextmv.local.application.poll")
+    def test_run_result_with_polling_success(self, mock_poll, mock_run_result, mock_run_metadata):
+        """Test successful run_result_with_polling call."""
 
         # Create mock run information with succeeded status
         mock_run_info = Mock(spec=RunInformation)
@@ -677,29 +569,25 @@ class TestApplicationLocalRunMethods(unittest.TestCase):
 
         # Mock the final result
         mock_result = Mock(spec=RunResult)
-        mock_local_run_result.return_value = mock_result
+        mock_run_result.return_value = mock_result
 
-        result = self.app.local_run_result_with_polling(self.test_run_id)
+        result = self.app.run_result_with_polling(self.test_run_id)
 
         # Verify poll was called
         mock_poll.assert_called_once()
 
         # Verify final result call
-        mock_local_run_result.assert_called_once_with(
+        mock_run_result.assert_called_once_with(
             run_id=self.test_run_id, run_information=mock_run_info, output_dir_path="."
         )
 
         self.assertEqual(result, mock_result)
 
-    @patch.object(Application, "local_run_metadata")
-    @patch.object(Application, "_Application__local_run_result")
-    @patch("nextmv.cloud.application.poll")
-    def test_local_run_result_with_polling_custom_options(
-        self, mock_poll, mock_local_run_result, mock_local_run_metadata
-    ):
-        """Test local_run_result_with_polling with custom polling options."""
-        from nextmv.cloud.application import PollingOptions
-        from nextmv.cloud.run import RunInformation, RunResult
+    @patch.object(Application, "run_metadata")
+    @patch.object(Application, "_Application__run_result")
+    @patch("nextmv.local.application.poll")
+    def test_run_result_with_polling_custom_options(self, mock_poll, mock_run_result, mock_run_metadata):
+        """Test run_result_with_polling with custom polling options."""
 
         # Create custom polling options
         custom_polling_options = PollingOptions(max_tries=10, max_duration=300)
@@ -710,9 +598,9 @@ class TestApplicationLocalRunMethods(unittest.TestCase):
         mock_run_info.metadata.status_v2 = StatusV2.succeeded
         mock_poll.return_value = mock_run_info
         mock_result = Mock(spec=RunResult)
-        mock_local_run_result.return_value = mock_result
+        mock_run_result.return_value = mock_result
 
-        self.app.local_run_result_with_polling(
+        self.app.run_result_with_polling(
             self.test_run_id, polling_options=custom_polling_options, output_dir_path="/custom/path"
         )
 
@@ -722,14 +610,13 @@ class TestApplicationLocalRunMethods(unittest.TestCase):
         self.assertEqual(call_args[1]["polling_options"], custom_polling_options)
 
         # Verify final result call with custom output path
-        mock_local_run_result.assert_called_once_with(
+        mock_run_result.assert_called_once_with(
             run_id=self.test_run_id, run_information=mock_run_info, output_dir_path="/custom/path"
         )
 
-    @patch.object(Application, "local_run_metadata")
-    def test_local_run_result_with_polling_status_check(self, mock_local_run_metadata):
+    @patch.object(Application, "run_metadata")
+    def test_run_result_with_polling_status_check(self, mock_run_metadata):
         """Test that polling function correctly checks for terminal statuses."""
-        from nextmv.cloud.run import RunInformation
 
         # Test different statuses
         test_cases = [
@@ -745,7 +632,7 @@ class TestApplicationLocalRunMethods(unittest.TestCase):
                 mock_run_info = Mock(spec=RunInformation)
                 mock_run_info.metadata = Mock()
                 mock_run_info.metadata.status_v2 = status
-                mock_local_run_metadata.return_value = mock_run_info
+                mock_run_metadata.return_value = mock_run_info
 
                 # We need to access the polling function that gets created inside the method
                 # This is a bit tricky, so we'll just test the logic conceptually
@@ -758,27 +645,26 @@ class TestApplicationLocalRunMethods(unittest.TestCase):
                     self.assertIsNone(result)
                     self.assertFalse(is_done)
 
-    # Tests for new_local_run_with_result
-    @patch.object(Application, "new_local_run")
-    @patch.object(Application, "local_run_result_with_polling")
-    def test_new_local_run_with_result_success(self, mock_polling_result, mock_new_local_run):
-        """Test successful new_local_run_with_result call."""
-        from nextmv.cloud.run import RunResult
+    # Tests for new_run_with_result
+    @patch.object(Application, "new_run")
+    @patch.object(Application, "run_result_with_polling")
+    def test_new_run_with_result_success(self, mock_polling_result, mock_new_run):
+        """Test successful new_run_with_result call."""
 
         # Mock responses
-        mock_new_local_run.return_value = "new-run-id"
+        mock_new_run.return_value = "new-run-id"
         mock_result = Mock(spec=RunResult)
         mock_polling_result.return_value = mock_result
 
         test_input = {"vehicles": [{"id": "v1"}]}
         test_options = {"duration": "10s"}
 
-        result = self.app.new_local_run_with_result(
+        result = self.app.new_run_with_result(
             input=test_input, name="Test Run", description="Test Description", run_options=test_options
         )
 
-        # Verify new_local_run was called correctly
-        mock_new_local_run.assert_called_once_with(
+        # Verify new_run was called correctly
+        mock_new_run.assert_called_once_with(
             input=test_input,
             name="Test Run",
             description="Test Description",
@@ -790,20 +676,18 @@ class TestApplicationLocalRunMethods(unittest.TestCase):
 
         # Verify polling was called correctly
         mock_polling_result.assert_called_once_with(
-            run_id="new-run-id", polling_options=_DEFAULT_POLLING_OPTIONS, output_dir_path="."
+            run_id="new-run-id", polling_options=DEFAULT_POLLING_OPTIONS, output_dir_path="."
         )
 
         self.assertEqual(result, mock_result)
 
-    @patch.object(Application, "new_local_run")
-    @patch.object(Application, "local_run_result_with_polling")
-    def test_new_local_run_with_result_all_parameters(self, mock_polling_result, mock_new_local_run):
-        """Test new_local_run_with_result with all parameters."""
-        from nextmv.cloud.application import PollingOptions
-        from nextmv.cloud.run import RunConfiguration, RunResult
+    @patch.object(Application, "new_run")
+    @patch.object(Application, "run_result_with_polling")
+    def test_new_run_with_result_all_parameters(self, mock_polling_result, mock_new_run):
+        """Test new_run_with_result with all parameters."""
 
         # Mock responses
-        mock_new_local_run.return_value = "comprehensive-run-id"
+        mock_new_run.return_value = "comprehensive-run-id"
         mock_result = Mock(spec=RunResult)
         mock_polling_result.return_value = mock_result
 
@@ -816,7 +700,7 @@ class TestApplicationLocalRunMethods(unittest.TestCase):
         test_input_dir = "/test/input/dir"
         test_output_dir = "/test/output/dir"
 
-        result = self.app.new_local_run_with_result(
+        result = self.app.new_run_with_result(
             input=test_input,
             name="Comprehensive Test",
             description="Full parameter test",
@@ -828,8 +712,8 @@ class TestApplicationLocalRunMethods(unittest.TestCase):
             output_dir_path=test_output_dir,
         )
 
-        # Verify new_local_run was called with all parameters
-        mock_new_local_run.assert_called_once_with(
+        # Verify new_run was called with all parameters
+        mock_new_run.assert_called_once_with(
             input=test_input,
             name="Comprehensive Test",
             description="Full parameter test",
@@ -846,50 +730,49 @@ class TestApplicationLocalRunMethods(unittest.TestCase):
 
         self.assertEqual(result, mock_result)
 
-    @patch.object(Application, "new_local_run")
-    @patch.object(Application, "local_run_result_with_polling")
-    def test_new_local_run_with_result_propagates_new_local_run_errors(self, mock_polling_result, mock_new_local_run):
-        """Test that new_local_run_with_result propagates errors from new_local_run."""
-        # Make new_local_run raise an error
-        mock_new_local_run.side_effect = ValueError("New local run error")
+    @patch.object(Application, "new_run")
+    @patch.object(Application, "run_result_with_polling")
+    def test_new_run_with_result_propagates_new_run_errors(self, mock_polling_result, mock_new_run):
+        """Test that new_run_with_result propagates errors from new_run."""
+        # Make new_run raise an error
+        mock_new_run.side_effect = ValueError("New local run error")
 
         with self.assertRaises(ValueError) as context:
-            self.app.new_local_run_with_result(input={"test": "data"})
+            self.app.new_run_with_result(input={"test": "data"})
 
         self.assertEqual(str(context.exception), "New local run error")
         # Verify polling was never called
         mock_polling_result.assert_not_called()
 
-    @patch.object(Application, "new_local_run")
-    @patch.object(Application, "local_run_result_with_polling")
-    def test_new_local_run_with_result_propagates_polling_errors(self, mock_polling_result, mock_new_local_run):
-        """Test that new_local_run_with_result propagates errors from polling."""
-        # Mock successful new_local_run but failing polling
-        mock_new_local_run.return_value = "error-run-id"
+    @patch.object(Application, "new_run")
+    @patch.object(Application, "run_result_with_polling")
+    def test_new_run_with_result_propagates_polling_errors(self, mock_polling_result, mock_new_run):
+        """Test that new_run_with_result propagates errors from polling."""
+        # Mock successful new_run but failing polling
+        mock_new_run.return_value = "error-run-id"
         mock_polling_result.side_effect = TimeoutError("Polling timeout")
 
         with self.assertRaises(TimeoutError) as context:
-            self.app.new_local_run_with_result(input={"test": "data"})
+            self.app.new_run_with_result(input={"test": "data"})
 
         self.assertEqual(str(context.exception), "Polling timeout")
-        # Verify new_local_run was called
-        mock_new_local_run.assert_called_once()
+        # Verify new_run was called
+        mock_new_run.assert_called_once()
 
-    @patch.object(Application, "new_local_run")
-    @patch.object(Application, "local_run_result_with_polling")
-    def test_new_local_run_with_result_minimal_parameters(self, mock_polling_result, mock_new_local_run):
-        """Test new_local_run_with_result with minimal parameters."""
-        from nextmv.cloud.run import RunResult
+    @patch.object(Application, "new_run")
+    @patch.object(Application, "run_result_with_polling")
+    def test_new_run_with_result_minimal_parameters(self, mock_polling_result, mock_new_run):
+        """Test new_run_with_result with minimal parameters."""
 
         # Mock responses
-        mock_new_local_run.return_value = "minimal-run-id"
+        mock_new_run.return_value = "minimal-run-id"
         mock_result = Mock(spec=RunResult)
         mock_polling_result.return_value = mock_result
 
-        result = self.app.new_local_run_with_result(input={"minimal": "test"})
+        result = self.app.new_run_with_result(input={"minimal": "test"})
 
         # Verify default parameters are used
-        mock_new_local_run.assert_called_once_with(
+        mock_new_run.assert_called_once_with(
             input={"minimal": "test"},
             name=None,
             description=None,
@@ -900,7 +783,7 @@ class TestApplicationLocalRunMethods(unittest.TestCase):
         )
 
         mock_polling_result.assert_called_once_with(
-            run_id="minimal-run-id", polling_options=_DEFAULT_POLLING_OPTIONS, output_dir_path="."
+            run_id="minimal-run-id", polling_options=DEFAULT_POLLING_OPTIONS, output_dir_path="."
         )
 
         self.assertEqual(result, mock_result)
